+++ CHANGELOG.md
@@ -1,9 +1,22 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2017-12-18
+
+### Added
+
+- The function `run()` accepts a callback as first argument, executed when the server is started.
+
+### Fixed
+
+- Use `127.0.0.1` instead `localhost` as default host.
+
 ## 1.0.0 - 2017-11-18
 
-First version with basic features
+First version with basic features.
+
+[1.1.0]: https://github.com/oscarotero/php-server-manager/compare/v1.0.0...v1.1.0